+++ libs/data-transformer.configurator.table/jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  name: 'data-transformer.configurator.table',
+  preset: '../../jest.config.js',
+  coverageDirectory: '../../coverage/libs/data-transformer.configurator.table',
+  snapshotSerializers: [
+    'jest-preset-angular/build/AngularNoNgAttributesSnapshotSerializer.js',
+    'jest-preset-angular/build/AngularSnapshotSerializer.js',
+    'jest-preset-angular/build/HTMLCommentSerializer.js',
+  ],
+};