+++ libs/data-transformer.configurator.table/src/lib/types.ts
@@ -1,0 +1,7 @@
+import { TableDataTransformerConfiguratorService } from './table-data-transformer-configurator.service';
+
+declare module '@spryker/data-transformer' {
+  interface DataTransformerConfiguratorRegistry {
+    table: TableDataTransformerConfiguratorService;
+  }
+}