+++ libs/styles/CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-next.30](https://github.com/spryker/ui-components/compare/@spryker/styles@0.2.0-next.29...@spryker/styles@0.2.0-next.30) (2021-02-24)
+
+
+### Features
+
+* **input.password:** add lib ([8deba36](https://github.com/spryker/ui-components/commit/8deba365be7c33fef72d8e4878a3d22536e76320))
+
+
+
+
+
+# [0.2.0-next.29](https://github.com/spryker/ui-components/compare/@spryker/styles@0.2.0-next.28...@spryker/styles@0.2.0-next.29) (2021-02-22)
+
+
+### Bug Fixes
+
+* **user-menu:** fix format ([6374b53](https://github.com/spryker/ui-components/commit/6374b53ebf040c1b3bb05bdc6f8a71dab768a9fd))
+* **user-menu:** refactoring after AA review ([1c382f3](https://github.com/spryker/ui-components/commit/1c382f34892d574b885c60f0d82bca85bafad7c5))
+
+
+### Features
+
+* **user-menu:** add components ([7e37ab2](https://github.com/spryker/ui-components/commit/7e37ab2097a943fc4e2fde0e5e46f80651a3ca7c))
+* **user-menu:** add lib ([866609b](https://github.com/spryker/ui-components/commit/866609be9cb90224aa09fce773dae7f194f847cf))
+
+
+
+
+
 # [0.2.0-next.28](https://github.com/spryker/ui-components/compare/@spryker/styles@0.2.0-next.27...@spryker/styles@0.2.0-next.28) (2021-02-01)
 
 **Note:** Version bump only for package @spryker/styles